tomcat_scanner/requirements.txt
python-nmap==0.6.1
netaddr==0.7.19
requests==2.18.4
